pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "thesis project"
version = "0.1.0"
description = ""
authors = ["Ariel Ebersberger <ariel@ebersberger.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
minigrid = "^2.3.1"
wandb = "^0.18.3"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
coloredlogs = "^15.0.1"
tqdm = "^4.66.5"
joblib = "^1.4.2"
annotated-types = "^0.7.0"
timm = "^1.0.11"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.3.3"
jupyter = "^1.1.1"

[tool.poetry.group.cpu]
optional = true

[tool.poetry.group.cpu.dependencies]
torch = { version = "^2.4.1", source = "torch_cpu", markers = "extra=='cpu' and extra!='gpu'" }

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch = { version = "^2.4.1", source = "torch_cuda", markers = "extra=='gpu' and extra!='cpu'" }

[[tool.poetry.source]]
name = "torch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "torch_cuda"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:wandb.*",
    "ignore::DeprecationWarning:pygame.*"
]